fastapi/requirements.txt
fastapi==0.95.1
requests_oauthlib==1.3.1
uvicorn==0.17.6
openai
pysondb
linkedin_api